refactor: Move package service to main module
This drasticlly improves load times on dialog modules as they
no longer make unnecessary network calls
