+++ src/app/core/core.module.ts
@@ -3,7 +3,6 @@ import { ElectronService } from './services/electron.service';
 import { ThemeService } from './services/theme.service';
 import { ThunderstoreService } from './services/thunderstore.service';
 import { PreferencesService } from './services/preferences.service';
-import { PackageService } from './services/package.service';
 import { DownloadService } from './services/download.service';
 import { DatabaseService } from './services/database.service';
 
@@ -15,7 +14,6 @@ import { DatabaseService } from './services/database.service';
     ThemeService,
     ThunderstoreService,
     PreferencesService,
-    PackageService,
     DownloadService,
     DatabaseService
   ]
