+++ src/app/main/package-selection/package-details/package-details.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Package, PackageVersion } from '../../../core/models/package.model';
 import { PreferencesService } from '../../../core/services/preferences.service';
 import { Subscription } from 'rxjs';
-import { PackageService } from '../../../core/services/package.service';
+import { PackageService } from '../../services/package.service';
 import { ElectronService } from '../../../core/services/electron.service';
 
 @Component({
