+++ src/app/main/package-selection/packages-page/step-one/step-one.component.spec.ts
@@ -4,7 +4,7 @@ import { StepOneComponent } from './step-one.component';
 import { MaterialModule } from '../../../../shared/material.module';
 import { ReactiveFormsModule } from '@angular/forms';
 import { Component, Input } from '@angular/core';
-import { PackageService } from '../../../../core/services/package.service';
+import { PackageService } from '../../../services/package.service';
 import { MockPackageService } from '../../../../core/services/mocks.spec';
 
 @Component({
