+++ src/app/main/package-selection/packages-page/step-three/step-three.component.spec.ts
@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { StepThreeComponent } from './step-three.component';
 import { MaterialModule } from '../../../../shared/material.module';
 import { ReactiveFormsModule } from '@angular/forms';
-import { PackageService } from '../../../../core/services/package.service';
+import { PackageService } from '../../../services/package.service';
 import { MockPackageService } from '../../../../core/services/mocks.spec';
 import { ScrollToModule } from '@nicky-lenaers/ngx-scroll-to';
